fix timing side-channel and add 503 maintenance redirect
